+++ collector-desktop/src/collector/desktop/tests/albumitems/UpdateAlbumItemTests.java
@@ -202,7 +202,7 @@ public class UpdateAlbumItemTests {
 		final String albumName = "Books";
 
 		AlbumItem referenceAlbumItem = createReferenceAlbumItem(albumName);
-		int newAlbumID = DatabaseWrapper.addNewAlbumItem(referenceAlbumItem, false, false);
+		long newAlbumID = DatabaseWrapper.addNewAlbumItem(referenceAlbumItem, false, false);
 	}
 
 	private AlbumItem createReferenceAlbumItem(String albumName) {
